+++ src/BackupTrait.php
@@ -17,7 +17,6 @@ use Cake\Core\Configure;
 use Cake\Datasource\ConnectionInterface;
 use Cake\Datasource\ConnectionManager;
 use InvalidArgumentException;
-use RuntimeException;
 use Symfony\Component\Filesystem\Filesystem;
 
 /**
@@ -30,32 +29,20 @@ trait BackupTrait
      * @since 2.4.0
      * @var array
      */
-    private static $validExtensions = ['sql.bz2' => 'bzip2', 'sql.gz' => 'gzip', 'sql' => false];
+    protected static $validExtensions = ['sql.bz2' => 'bzip2', 'sql.gz' => 'gzip', 'sql' => false];
 
     /**
      * Returns an absolute path
      * @param string $path Relative or absolute path
      * @return string
-     * @uses getTarget()
      */
     public function getAbsolutePath($path)
     {
-        return (new Filesystem())->isAbsolutePath($path) ? $path : add_slash_term($this->getTarget()) . $path;
-    }
+        if (!(new Filesystem())->isAbsolutePath($path)) {
+            return add_slash_term(Configure::read('DatabaseBackup.target')) . $path;
+        }
 
-    /**
-     * Gets a binary path
-     * @param string $name Binary name
-     * @return string
-     * @since 2.0.0
-     * @throws \RuntimeException
-     */
-    public function getBinary($name)
-    {
-        $binary = Configure::read('DatabaseBackup.binaries.' . $name);
-        is_true_or_fail($binary, __d('database_backup', 'Binary for `{0}` could not be found. You have to set its path manually', $name), RuntimeException::class);
-
-        return $binary;
+        return $path;
     }
 
     /**
@@ -92,12 +79,11 @@ trait BackupTrait
      * @since 2.0.0
      * @throws \InvalidArgumentException
      * @uses getConnection()
-     * @uses getDriverName()
      */
     public function getDriver(ConnectionInterface $connection = null)
     {
         $connection = $connection ?: $this->getConnection();
-        $className = $this->getDriverName($connection);
+        $className = get_class_short_name($connection->getDriver());
         $driver = App::classname(sprintf('%s.%s', 'DatabaseBackup', $className), 'Driver');
         is_true_or_fail(
             $driver,
@@ -109,40 +95,17 @@ trait BackupTrait
     }
 
     /**
-     * Gets the driver name, according to the database engine
-     * @param \Cake\Datasource\ConnectionInterface|null $connection A connection object
-     * @return string The driver name
-     * @since 2.6.2
-     * @uses getConnection()
-     */
-    public function getDriverName(ConnectionInterface $connection = null)
-    {
-        $connection = $connection ?: $this->getConnection();
-
-        return get_class_short_name($connection->getDriver());
-    }
-
-    /**
      * Returns the extension from a filename
      * @param string $filename Filename
      * @return string|null Extension or `null` if the extension is not found or
      *  if is an invalid extension
-     * @uses getValidExtensions()
+     * @uses $validExtensions
      */
     public function getExtension($filename)
     {
         $extension = get_extension($filename);
 
-        return in_array($extension, $this->getValidExtensions()) ? $extension : null;
-    }
-
-    /**
-     * Returns the target path
-     * @return string
-     */
-    public function getTarget()
-    {
-        return Configure::read('DatabaseBackup.target');
+        return in_array($extension, array_keys(self::$validExtensions)) ? $extension : null;
     }
 
     /**
@@ -154,16 +117,5 @@ trait BackupTrait
     public function getValidCompressions()
     {
         return array_filter(self::$validExtensions);
-    }
-
-    /**
-     * Returns all valid extensions
-     * @return array
-     * @since 2.4.0
-     * @uses $validExtensions
-     */
-    public function getValidExtensions()
-    {
-        return array_keys(self::$validExtensions);
     }
 }