+++ src/Driver/Mysql.php
@@ -30,6 +30,7 @@ class Mysql extends Driver
     /**
      * Gets the executable command to export the database
      * @return string
+     * @uses getBinary()
      * @uses getConfig()
      * @uses $auth
      */
@@ -46,6 +47,7 @@ class Mysql extends Driver
     /**
      * Gets the executable command to import the database
      * @return string
+     * @uses getBinary()
      * @uses getConfig()
      * @uses $auth
      */