+++ src/Utility/BackupManager.php
@@ -68,7 +68,7 @@ class BackupManager
      */
     public function index()
     {
-        $finder = (new Finder())->files()->name('/\.sql(\.(gz|bz2))?$/')->in($this->getTarget());
+        $finder = (new Finder())->files()->name('/\.sql(\.(gz|bz2))?$/')->in(Configure::read('DatabaseBackup.target'));
 
         return collection($finder)->map(function (SplFileInfo $file) {
             return new Entity([